sales_ai/requirements.txt
Flask==3.1.0
openai==1.107.0
google-genai==1.30.0
python-dotenv==1.0.1
